requirements.txt
pandas
openreview-py
pytest
ortools
matplotlib
jupyter
tqdm